feat: add Vectors component for enhanced visual elements in login and registration pages, update Arabic translations, and improve error logging
