+++ translation/ar.json
@@ -27,7 +27,7 @@
   },
   "Login": {
     "Welcome": "مرحبا بك",
-    "We are glad to see you back with us": "تسجيل الدخول",
+    "We are glad to see you back with us": "سعداء برجوعك إلينا",
     "Email": "البريد الإلكتروني",
     "Password": "كلمة المرور",
     "Forget Your Password?": "نسيت كلمة المرور؟",
